for correction

+++ src/airport_type.hpp
@@ -30,8 +30,8 @@ public:
         return std::vector<Terminal> { terminal_pos.begin(), terminal_pos.end() };
     }
 
-    WaypointQueue air_to_terminal(const Point3D& offset, const size_t runway_num,
-                                  const size_t terminal_num) const
+    WaypointQueue air_to_terminal(const Point3D& offset, const int runway_num,
+                                  const int terminal_num) const
     {
         const Runway& runway = runways.at(runway_num);
 
@@ -54,8 +54,8 @@ public:
         return result;
     }
 
-    WaypointQueue terminal_to_air(const Point3D& offset, const size_t runway_num,
-                                  const size_t terminal_num) const
+    WaypointQueue terminal_to_air(const Point3D& offset, const int runway_num,
+                                  const int terminal_num) const
     {
         const Runway& runway = runways.at(runway_num);
         const float angle    = (rand() % 1000) * 2 * 3.141592f / 1000.f; // random angle between 0 and 2pi
